add EditPost to TgBotActor, change messageId type to Int

+++ src/main/scala/com/github/awant/habrareader/models/Event.scala
@@ -6,7 +6,7 @@ import io.circe._
 import io.circe.syntax._
 import com.github.awant.habrareader.utils.DateUtils._
 
-case class Event(id: Long, chatId: Long, messageId: Long, postId: Long, update: Date)
+case class Event(id: Long, chatId: Long, messageId: Int, postId: Long, update: Date)
 
 object Event {
 
@@ -26,7 +26,7 @@ object Event {
     for {
       id <- c.get[Long]("id")
       chatId <- c.get[Long]("chatId")
-      messageId <- c.get[Long]("messageId")
+      messageId <- c.get[Int]("messageId")
       postId <- c.get[Long]("postId")
       update <- c.get[Date]("update")
     } yield Event(id, chatId, messageId, postId, update)
